rename top level package scitt -> datatrails_scitt_samples
