+++ .github/workflows/registration-demo.yml
@@ -38,8 +38,8 @@ jobs:
         shell: bash
       - name: Generate ephemeral issuer key
         run: |
-          python3 -m scitt.scripts.generate_example_key.py \
-            --signing-key-file-path scitt-signing-key.pem
+          python3 -m datatrails_scitt_samples.scripts.generate_example_key.py \
+            --signing-key-file-path datatrails_scitt_samples-signing-key.pem
 
       - name: Create statement
         env:
@@ -52,8 +52,8 @@ jobs:
           # NOTE: We are reling on github's ui to sanitize the inputs
 
           echo ${{ inputs.payload }} > payload.json
-          python3 -m scitt.scripts.create_signed_statement \
-            --signing-key-file-path scitt-signing-key.pem \
+          python3 -m datatrails_scitt_samples.scripts.create_signed_statement \
+            --signing-key-file-path datatrails_scitt_samples-signing-key.pem \
             --payload-file payload.json \
             --content-type ${{ inputs.content_type }} \
             --subject ${{ inputs.subject }} \
@@ -68,7 +68,7 @@ jobs:
 
         run: |
 
-          python3 -m scitt.scripts.register_signed_statement \
+          python3 -m datatrails_scitt_samples.scripts.register_signed_statement \
             --signed-statement-file signed-statement.cbor \
             --output-file transparent-statement.cbor \
             --output-receipt-file statement-receipt.cbor
