+++ datatrails_scitt_samples/.gitignore
