+++ datatrails_scitt_samples/artifacts/thedroid.json
