+++ datatrails_scitt_samples/datatrails/.gitignore
