+++ datatrails_scitt_samples/mmriver/algorithms.py
