+++ datatrails_scitt_samples/scripts/.gitignore
