+++ datatrails_scitt_samples/scripts/create_signed_statement.py
@@ -3,8 +3,8 @@
 import sys
 import argparse
 
-from scitt.scripts.fileaccess import open_payload, open_signing_key
-from scitt.statement_creation import create_signed_statement
+from datatrails_scitt_samples.scripts.fileaccess import open_payload, open_signing_key
+from datatrails_scitt_samples.statement_creation import create_signed_statement
 
 
 def main(args=None):
