+++ datatrails_scitt_samples/scripts/fileaccess.py
