+++ datatrails_scitt_samples/scripts/generate_example_key.py
