+++ datatrails_scitt_samples/scripts/verify_receipt.py.disabled
