+++ unittests/test_verify_receipt_signature.py
@@ -4,8 +4,8 @@ Known Answer Test (KAT) unit tests for verifying a receipt
 
 import unittest
 
-# from scitt.cose_receipt_verification import verify_receipt_mmriver
-# from scitt.scripts.fileaccess import read_cbor_file
+# from datatrails_scitt_samples.cose_receipt_verification import verify_receipt_mmriver
+# from datatrails_scitt_samples.scripts.fileaccess import read_cbor_file
 # from .constants import KNOWN_RECEIPT_FILE
